requirements.txt
python-dotenv==0.19.2
requests==2.27.1
commentjson==0.9.0